fix: improve upgrade script
